+++ app/Http/Controllers/ProductionController.php
@@ -30,9 +30,9 @@ class ProductionController extends Controller
         $multiFactory = MultiFactories::find(request('multiFactory'));
         $choices = !empty(request('choices',[])) ? collect(request('choices',(object) []))->reject(fn($name) => r($name)->isDefault())->all() : null;
         $even = request('even') ? 1 : 0;
+        $speedLimit = request('speedLimit', 'both');
 
-
-        return compact('products','recipes','favorites','factory','multiFactory','choices','even');
+        return compact('products','recipes','favorites','factory','multiFactory','choices','even','speedLimit');
     }
 
     public function index()